requirements.txt
mysql-connector-python
pillow
reportlab
 